Spostati alcuni metodi.
