+++ BusinessLayer/BL_LinkManagement.cs
@@ -1,0 +1,14 @@
+﻿using SchoolGrades.DbClasses;
+using System;
+using System.Collections.Generic;
+using System.Data;
+using System.Data.Common;
+using System.Text;
+
+namespace SchoolGrades
+{
+    internal partial class BusinessLayer
+    {
+        
+    }
+}
