+++ BusinessLayer/BusinessLayer.projitems
@@ -9,9 +9,12 @@
     <Import_RootNamespace>BusinessLayer</Import_RootNamespace>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)BL_AnnotationManagement.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)BL_AnswerManagement.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)BL_ClassManagement.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)BL_ConstructorsAndGeneralMethods.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)BL_GradesManagement.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)BL_LinkManagement.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)BL_StudentManagement.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)BL_UserManagement.cs" />
   </ItemGroup>
