+++ DataLayer/DL_AnswerManagement.cs
@@ -1,0 +1,71 @@
+﻿using SchoolGrades.DbClasses;
+using System;
+using System.Collections.Generic;
+using System.Data.Common;
+using System.Text;
+
+namespace SchoolGrades
+{
+    internal partial class DataLayer
+    {
+
+        internal void AddAnswerToQuestion(int? idQuestion, int? idAnswer)
+        {
+            using (DbConnection conn = Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "UPDATE Answers" +
+                    " SET idAnswer=" + idAnswer +
+                    " WHERE idQuestion =" + idQuestion +
+                    ";";
+                cmd.ExecuteNonQuery();
+                cmd.Dispose();
+            }
+        }
+
+        internal int CreateAnswer(Answer currentAnswer)
+        {
+            // trova una chiave da assegnare alla nuova domanda
+            int codice = NextKey("Answers", "idAnswer");
+            using (DbConnection conn = Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "INSERT INTO Answers" +
+                    " (idAnswer,idQuestion,showingOrder,text,errorCost,isCorrect,isOpenAnswer)" +
+                    " Values (" + codice +
+                    "," + SqlVal.SqlInt(currentAnswer.IdQuestion) +
+                    "," + SqlVal.SqlInt(currentAnswer.ShowingOrder) +
+                    ",'" + SqlVal.SqlString(currentAnswer.Text) + "'" +
+                    "," + SqlVal.SqlDouble(currentAnswer.ErrorCost) +
+                    "," + SqlVal.SqlBool(currentAnswer.IsCorrect) +
+                    "," + SqlVal.SqlBool(currentAnswer.IsOpenAnswer) +
+                    ");";
+                cmd.ExecuteNonQuery();
+
+                cmd.Dispose();
+            }
+            return codice;
+        }
+
+        internal void SaveAnswer(Answer currentAnswer)
+        {
+            using (DbConnection conn = Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "UPDATE Answers" +
+                    " SET idAnswer=" + currentAnswer.IdAnswer + "," +
+                    " idQuestion=" + currentAnswer.IdQuestion + "," +
+                    " isCorrect='" + SqlVal.SqlBool(currentAnswer.IsCorrect) + "'," +
+                    " isOpenAnswer='" + SqlVal.SqlBool(currentAnswer.IsOpenAnswer) + "'," +
+                    " Text='" + SqlVal.SqlString(currentAnswer.Text) + "'," +
+                    " errorCost=" + SqlVal.SqlDouble(currentAnswer.ErrorCost.ToString()) + "" +
+                    " WHERE idAnswer = " + currentAnswer.IdAnswer +
+                    ";";
+                cmd.ExecuteNonQuery();
+
+                cmd.Dispose();
+            }
+        }
+
+    }
+}
